refactor(RacingCarServiceTest): 서비스 최종 우승자 구할 테스트
